requirements.txt
flask
flask-SocketIO
requests
python-dotenv
gunicorn
eventlet